fixed warnings (#1357)
* fixed warnings


* Added cosmetic suggestions


* Refs 9193. Uncrustify


* Refs 9193. Fixed additional warnings.

+++ include/fastdds/dds/core/policy/QosPolicies.hpp
@@ -787,7 +787,7 @@ public:
         , ResourceLimitedOctetVector()
     {
         assign(data.begin(), data.end());
-        length = (size() + 7) & ~3;
+        length = static_cast<uint16_t>((size() + 7u) & ~3u);
     }
 
     virtual RTPS_DllAPI ~GenericDataQosPolicy() = default;
@@ -811,7 +811,7 @@ public:
             //If the object is size limited, already has max_size() allocated
             //assign() will always stop copying when reaching max_size()
             assign(b.begin(), b.end());
-            length = (size() + 7) & ~3;
+            length = static_cast<uint16_t>((size() + 7u) & ~3u);
             hasChanged = true;
         }
         return *this;
@@ -917,7 +917,7 @@ public:
         if (collection_ != vec)
         {
             assign(vec.begin(), vec.end());
-            length = (size() + 7) & ~3;
+            length = static_cast<uint16_t>((size() + 7u) & ~3u);
             hasChanged = true;
         }
     }
@@ -1332,7 +1332,7 @@ public:
         {
             //Size of the element (with alignment)
             uint32_t size = *(uint32_t*)ptr_;
-            ptr_ += (4 + ((size + 3) & ~3));
+            ptr_ += (4u + ((size + 3u) & ~3u));
             value_ = Partition_t(ptr_);
         }
 
@@ -1484,7 +1484,7 @@ public:
     {
         //Realloc if needed;
         uint32_t size = (uint32_t)strlen(name) + 1;
-        uint32_t alignment = ((size + 3) & ~3) - size;
+        uint32_t alignment = ((size + 3u) & ~3u) - size;
 
         if (max_size_ != 0 && (partitions_.max_size < partitions_.length +
                 size + alignment + 4))
@@ -2659,7 +2659,7 @@ public:
     }
 
     //!User defined transports to use alongside or in place of builtins.
-    std::vector<std::shared_ptr<fastdds::rtps::TransportDescriptorInterface> > user_transports;
+    std::vector<std::shared_ptr<fastdds::rtps::TransportDescriptorInterface>> user_transports;
 
     //!Set as false to disable the default UDPv4 implementation. <br> By default, true.
     bool use_builtin_transports;
